requirements.txt
Django==4.0.4

# Django related packages
django-braces==1.15.0
django-simple-menu==1.2.1
social-auth-app-django==5.0.0
django-rosetta==0.9.8
social-auth-core==4.2.0
django-tqdm==1.1.0
django-modeladmin-reorder==0.3.1
# Django debug
django-debug-toolbar
django-debug-toolbar-template-timings

mysqlclient==2.1.0
sentry-sdk==1.5.10
PyGithub==1.55
gunicorn==20.1.0
